It's almost there + gitignore
The hash table is almost building the faces in a proper way. I thing I
need to threat the colisions a bit better. Lets consider some equal
indexes in the vector and proceed by this way.

Hey.. The gitignore changed a bit too. Vim guys suffer with those *.swp
files...

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
    modified:   .gitignore
    modified:   unstr_ds.f90