Now the colision counter is working.
I maneged to artificially make the colision counter work. Now it's just
the case to use it in a mesh that presents colision.

The problem now is that my mesh converter is not working with any kind
of mesh that represents an external flow. I have to debug that guy and
it won't be easy.

On branch deal_colision
Changes to be committed:
    modified:   unstr_ds.f90