Fixed the problem with faces creation.
During the faces creation, I had some problems during allocation. The
problem was caused by the last element that needs one more face than the
rest of the faces.

To solve this I did a calloc like subroutine that re-dimension the face
vector. Until now, it is solved !

On branch deal_colision
Changes to be committed:
    modified:   Makefile
    modified:   unstr_ds.f90