Forget the colision for a while.
There are so many more interesting things to do in the code then dealing
with colisions.

The best attempt until now is now implemented, linear probing.

The problem is that, when you a colision happens, the face(_,4) is not
filled properly. I think the rest of the vector is bad too.

On branch deal_colision
Changes to be committed:
    modified:   input.in
    modified:   nlb2d.f90
    modified:   preproc.f90